UBUNTU: add Breaks: against hardy lvm2
BugLink: http://bugs.launchpad.net/bugs/528155

Installing lucid kernel on to a Hardy system which uses lvm2 will result in
an unbootable system.  While this is not supported per-see it is possible
to get into this combination with an interrupted upgrade.  It is easy to
avoid by ensuring lvm2 gets upgraded first via a Breaks: directive.

Acked-by: Colin King <colin.king@canonical.com>
Acked-by: Amit Kucheria <amit.kucheria@canonical.com>
Acked-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>