(Upstream) dell-laptop: defer dell_rfkill_update to worker thread
dell_rfkill_update fires an SMI, which must occur on cpu 0. Thus, if called
on a different cpu, the task will block. Since it's called in hard irq
context, we must defer this to the worker thread.

It is potentially possible that two rfkill key events could be processed
before the work completes on the first. The second event is dropped with a
KERN_NOTICE message.

BugLink: http://bugs.launchpad.net/bugs/555261

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>