UBUNTU: SAUCE: ACPI: EC: Allow multibyte access to EC (v3)
Allow more than a single byte to be read or written to the EC.  This is
needed by some Dell BIOSs.  Do not allow access to exceed 64 bits as the
underlying algorith assumes it can do single aligned accesses of this size.

Based on the original version by Alexey Starikovskiy.

BugLink: http://bugs.launchpad.net/bugs/526354
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>